refactor: split trait_impl/verify_d.rs under 500-LoC cap
Extract the K=gamma verify per-layer forward loop (attention/SSM
dispatch + DFlash hidden capture) into a sibling
verify_d_kgamma_layers.rs, returning per-phase timing via a
KgammaLayerTiming struct so the caller's ATLAS_DFLASH_TIMING
aggregation is unaffected.
